feature: 스레드간 통신
