+++ client/src/launcher/online_data_temp.py
@@ -1,0 +1,11 @@
+from collections import deque
+
+
+class OnlineData:
+    def __init__(self):
+        self.message_queue = deque([])
+
+    def message_queue_gen(self):
+        while True:
+            if self.message_queue:
+                yield self.message_queue.popleft()
